back/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cors-headers==3.7.0
djangorestframework==3.12.2
pytz==2020.5
sqlparse==0.4.1
psycopg2
gunicorn==20.0.4
whitenoise==5.2.0
PyDriller~=1.15.5
python-dotenv~=0.15.0
Faker~=6.4.1
nose~=1.3.7